pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-storages"
description = 'FastAPI Storages'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["sqlalchemy", "django", "orm"]
authors = [
  { name = "Amin Alaee", email = "me@aminalaee.dev" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Database",
  "Topic :: Database :: Database Engines/Servers",
  "Topic :: Internet",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
  "Topic :: Internet :: WWW/HTTP",
  "Typing :: Typed",
]
dependencies = [
  "boto3~=1.25",
]
dynamic = ["version"]

[project.optional-dependencies]
full = [
  "Pillow~=9.4",
  "sqlalchemy>=1.4",
]

[project.urls]
Documentation = "https://github.com/aminalaee/fastapi-storages#readme"
Issues = "https://github.com/aminalaee/fastapi-storages/issues"
Source = "https://github.com/aminalaee/fastapi-storages"

[tool.hatch.version]
path = "fastapi_storages/__init__.py"

[tool.hatch.build.targets.wheel]
[tool.hatch.build.targets.sdist]
include = [
    "/fastapi_storages",
]

[tool.hatch.build]
exclude = [
  "tests/*",
]

[tool.hatch.envs.default]
dependencies = [
  "black==22.10.0",
  "build==0.9.0",
  "coverage==6.5.0",
  "moto==4.1.2",
  "mypy==0.982",
  "Pillow==9.4.0",
  "pytest==7.2.0",
  "ruff==0.0.237",
  "sqlalchemy>=1.4",
]

[tool.hatch.envs.default.scripts]
check = [
  "ruff .",
  "black --check .",
  "mypy fastapi_storages",
]
cov = [
  "coverage report --show-missing --skip-covered --fail-under=99",
  "coverage xml",
]
format = [
  "black .",
  "ruff --fix .",
]
test = "coverage run -a -m pytest {args}"

[tool.hatch.envs.docs]
dependencies = [
  "mkdocs-material==8.5.7",
  "mkdocs==1.4.1",
  "mkdocstrings[python]==0.18.1",
]

[tool.hatch.envs.docs.scripts]
build = "mkdocs build"
serve = "mkdocs serve --dev-addr localhost:8080"
deploy = "mkdocs gh-deploy --force"

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
no_implicit_optional = true

[tool.ruff]
select = ["E", "F", "I"]

[tool.coverage.run]
source_pkgs = [
  "fastapi_storages",
  "tests",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "pragma: nocover",
    "except NotImplementedError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
